# docs/swagger.yaml
basePath: /api
definitions:
...
  gin.H:
    additionalProperties: true
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Directories:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      disabled:
        type: boolean
      id:
        type: integer
      name:
        type: string
      subscriptions:
        items:
          $ref: '#/definitions/models.Subscriptions'
        type: array
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.DirectoriesUsers'
        type: array
      uuid:
        type: string
    type: object
  models.DirectoriesUsers:
    properties:
      directory:
        $ref: '#/definitions/models.Directories'
      directory_id:
        type: integer
      user_id:
        type: integer
      users:
        $ref: '#/definitions/models.Users'
    type: object
  models.Permissions:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.RolesPermissions'
        type: array
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  models.RBACResourcePools:
    properties:
      resource_pool:
        $ref: '#/definitions/models.ResourcePools'
      resource_pool_id:
        type: integer
      role:
        $ref: '#/definitions/models.Roles'
      role_id:
        type: integer
      user:
        $ref: '#/definitions/models.Users'
      user_id:
        type: integer
    type: object
  models.RBACSubscriptions:
    properties:
      role:
        $ref: '#/definitions/models.Roles'
      role_id:
        type: integer
      subscription:
        $ref: '#/definitions/models.Subscriptions'
      subscription_id:
        type: integer
      user:
        $ref: '#/definitions/models.Users'
      user_id:
        type: integer
    type: object
  models.ResourcePools:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      rbac:
        items:
          $ref: '#/definitions/models.RBACResourcePools'
        type: array
      resources:
        items:
          $ref: '#/definitions/models.Resources'
        type: array
      subscription:
        $ref: '#/definitions/models.Subscriptions'
      subscription_id:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  models.Resources:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      resource_pool:
        $ref: '#/definitions/models.ResourcePools'
      resource_pool_id:
        type: integer
      resource_status:
        $ref: '#/definitions/models.ResourcesStatus'
      resource_status_id:
        type: integer
      service:
        $ref: '#/definitions/models.ServiceVersions'
      service_id:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  models.ResourcesStatus:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      resources:
        items:
          $ref: '#/definitions/models.Resources'
        type: array
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  models.Roles:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.RolesPermissions'
        type: array
      rbac_resource_pools:
        items:
          $ref: '#/definitions/models.RBACResourcePools'
        type: array
      rbac_subscriptions:
        items:
          $ref: '#/definitions/models.RBACSubscriptions'
        type: array
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  models.RolesPermissions:
    properties:
      permission:
        $ref: '#/definitions/models.Permissions'
      permission_id:
        type: integer
      role:
        $ref: '#/definitions/models.Roles'
      role_id:
        type: integer
    type: object
  models.ServiceVersions:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      image:
        type: string
      resources:
        items:
          $ref: '#/definitions/models.Resources'
        type: array
      service:
        $ref: '#/definitions/models.Services'
      service_id:
        type: integer
      updated_at:
        type: string
      uuid:
        type: string
      version:
        type: string
    type: object
  models.Services:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
      versions:
        items:
          $ref: '#/definitions/models.ServiceVersions'
        type: array
    type: object
  models.Subscriptions:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      directory:
        $ref: '#/definitions/models.Directories'
      directory_id:
        type: integer
      disabled:
        type: boolean
      id:
        type: integer
      name:
        type: string
      rbac:
        items:
          $ref: '#/definitions/models.RBACSubscriptions'
        type: array
      resource_pools:
        items:
          $ref: '#/definitions/models.ResourcePools'
        type: array
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  models.Users:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      directories:
        items:
          $ref: '#/definitions/models.DirectoriesUsers'
        type: array
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      need_chpasswd:
        type: boolean
      passwd_checksum:
        type: string
      rbac_resource_pools:
        items:
          $ref: '#/definitions/models.RBACResourcePools'
        type: array
      rbac_subscriptions:
        items:
          $ref: '#/definitions/models.RBACSubscriptions'
        type: array
      temporary_password:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  types.ArrayResponse:
    properties:
      record_count:
        type: integer
      record_list: {}
    type: object
...
info:
  contact: {}
  description: This is a Cloud COCO API
  license:
    name: Apache License Version 2.0
    url: https://github.com/mrzack99s/cloud-coco
  title: Cloud COCO
  version: 0.1.0
paths:
...
  /resource-pools/create:
    post:
      consumes:
      - application/json
      description: Create ResourcePools
      operationId: create-resource-pools
      parameters:
      - description: Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.ResourcePools'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResourcePools'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Create ResourcePools
      tags:
      - ResourcePools
  /resource-pools/get-by-offset:
    get:
      description: Get ResourcePools by offset
      operationId: get-resource-pools-by-offset
      parameters:
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ArrayResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Get ResourcePools by offset
      tags:
      - ResourcePools
  /resource-pools/get/{uuid}:
    get:
      description: Get ResourcePools by uuid
      operationId: get-resource-pools-by-uuid
      parameters:
      - description: UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResourcePools'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Get ResourcePools by uuid
      tags:
      - ResourcePools
  /resource-pools/hard-delete/{uuid}:
    delete:
      description: Hard Delete ResourcePools
      operationId: hard-delete-resource-pools
      parameters:
      - description: ResourcePools UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Hard Delete ResourcePools
      tags:
      - ResourcePools
  /resource-pools/soft-delete/{uuid}:
    delete:
      description: Soft Delete ResourcePools
      operationId: soft-delete-resource-pools
      parameters:
      - description: ResourcePools UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Soft Delete ResourcePools
      tags:
      - ResourcePools
  /resource-pools/update:
    post:
      consumes:
      - application/json
      description: Update ResourcePools
      operationId: update-resource-pools
      parameters:
      - description: Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.ResourcePools'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResourcePools'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Update ResourcePools
      tags:
      - ResourcePools
  /resources/create:
    post:
      consumes:
      - application/json
      description: Create Resources
      operationId: create-resources
      parameters:
      - description: Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Resources'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resources'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Create Resources
      tags:
      - Resources
  /resources/get-by-offset:
    get:
      description: Get Resources by offset
      operationId: get-resources-by-offset
      parameters:
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ArrayResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Get Resources by offset
      tags:
      - Resources
  /resources/get/{uuid}:
    get:
      description: Get Resources by uuid
      operationId: get-resources-by-uuid
      parameters:
      - description: UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resources'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Get Resources by uuid
      tags:
      - Resources
  /resources/hard-delete/{uuid}:
    delete:
      description: Hard Delete Resources
      operationId: hard-delete-resources
      parameters:
      - description: Resources UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Hard Delete Resources
      tags:
      - Resources
  /resources/soft-delete/{uuid}:
    delete:
      description: Soft Delete Resources
      operationId: soft-delete-resources
      parameters:
      - description: Resources UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Soft Delete Resources
      tags:
      - Resources
  /resources/update:
    post:
      consumes:
      - application/json
      description: Update Resources
      operationId: update-resources
      parameters:
      - description: Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Resources'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resources'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Update Resources
      tags:
      - Resources
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-access-token
    type: apiKey
swagger: "2.0"